# Conflicts:
#	src/main/java/org/example/psychologicalcounseling/constant/ErrorConstant.java
#	src/main/java/org/example/psychologicalcounseling/controller/chat/ChatWithGPTController.java

+++ src/main/java/org/example/psychologicalcounseling/repository/UserRepository.java
@@ -6,9 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     //表格如同实体类User，uid是主键，主键为Long类型
     //继承了JpaRepository，自带对于Users表格的增删改查方法
 
     User findByEmail(String email);
+
 }
